prefetcher: cancel all prefetches for a given TLF
Previously, when changing a sync config, we would toggle the whole
prefetcher to make sure the old requests were canceled.  But this can
block for a really long time if there are a lot of requests
outstanding.  We were doing this to make sure that when a block no
longer needs to be prefetched (due to the config change), it wouldn't
continue to be prefetched.  But we can handle this a different way,
without all the blocking.

Instead, When a sync config is disabled, or a partial sync config is
changed, just cancel all the prefetches associated with that TLF ID.
It should be relatively rare, so that it doesn't matter much that we
have to loop through all the prefetches when it happens.

Do it synchronously so we can make sure all the requests are canceled
before the disk cache is scrubbed.

Issue: KBFS-4070